requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0.post1